Removes Dr.V from the game  (#1428)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
it was stated in dev chat that Dr.V is a self insert character and was
not authorized by a lore writer to be added to the game


![image](https://github.com/user-attachments/assets/5d352917-7e8d-40d6-b74b-200ca739d259)

this is a quick, tested PR that removes Dr.V from the desc text along
with removing the OOC refrences in desc text to preserve immersion

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

By removing a self insert character that was not authorized by lore
writers, we preserve the codebase guidelines

by removing OOC text from description, we preserve immersion.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->