Fixes steel flail icon (#1430)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Don't know if anyone has reported or noticed this yet but, while doing
things on the map editor i noticed that someone accidentally changed the
icon_state on the flail to "steel flail", which in fact does not exist,
this PR fixes this and gives the steel flail an icon again.
Before:
![Captura de pantalla 2024-11-04
142244](https://github.com/user-attachments/assets/ac1c74cf-8df2-478d-81af-8f6c1967fcf0)
After:
![Captura de pantalla 2024-11-04
142207](https://github.com/user-attachments/assets/ea906dc3-22fe-461b-a198-01aa47d5c800)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Fixes a bug, aka, makes the game better.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->