Hermes & Stockpile QoL changes (#1410)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- Ports
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/256
- Ports the following from
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/340:
- > Adds COMPACT MODE to the stockpile, vomitarium, and steward's
console. You will NOT need to scroll anymore.

Proof of testing:

![image](https://github.com/user-attachments/assets/1e7f5c13-39ff-4c8c-9ee4-dd9f918ff55a)

(Not the actual directories, this was tested with FASTLOAD)

![image](https://github.com/user-attachments/assets/aabe09ef-1c28-46f7-8c73-b2dd4ae32ffa)


## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Basically just raw QoL.