Clinic Removal (#1435)
## About The Pull Request
To be clear, there's probably vestigial elements in the code still, but
I 'unno about that. A quick search got everything.

In any case, this returns the map's old layout before the Clinic, for
the most part.
Removes the Clinic Physician. Gives their in game material to the Court
Physician.
Adjusts the Town Physician's building to keep up.

## Why It's Good For The Game
The Clinic took emphasis away from the Court Physician, alongside the
Town's Doctors, before throwing them into a role that was overall
redundant. I _**PERSONALLY**_ feel like it was added for the sake of it,
not that there's anything wrong with that, though with no consideration
to anything else. Especially as the individual who made the PR
specifically called out Court Physician in a negative light, pushing the
changes more into a spite direction. I've had a number of folks approach
me for this PR in particular, after I proposed a revert on the Clinic
mapping and roles.