Mercenary Guild Updates 2.0 (#1438)
## About The Pull Request
As for the main bulk of the PR, the guild hall now receives a more
welcoming entrance, a garden, a larder(?) and a visual facelift to some
of its areas. The second floor also has a roofing lip now, as to prevent
kobolds from climbing up inside and stealing the jupons roundstart.
Additionally, the Guild Clerk's stats have been touched up, to include
climbing, swimming and a (very)fat coin purse for paying out contracts.

## Why It's Good For The Game
Just additional facelifts and QoL content to the guild, which had been
sorely lacking prior to the previous PR. This just touches it up
further.