Adds Unfinished Ratwood Map (#1448)
# Adds Incomplete map
Adding unfinished map to codebase

## Permissions

This map is work in progress. All PR's related to this map until it is
finished must be approved by its creator (myself in this instance).

## Contribution
Any PR's made to this map must only involve haeangongguk.dmm


This map is being added to the codebase in its unfinished state due to
the following
- In the event of death of creator
- To allow all contributors to provide possible suggestions through PR
or messaging creator of map on Discord
- To give bored people a look at a future map
- Creator of map has hit a wall in terms of structure layouts (I make
spaceships, not medieval cities).