Alternative Latejoin location: The Carriage (#1449)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

- Ports
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/598.

Adds a new location for latejoining, the carriage. It is located south
west of the town, in that sewer/swamp area previously unused. It's a 50
50 whether you spawn by the boat or here.

![image](https://github.com/user-attachments/assets/9d5d6224-7748-4271-94d0-a7151c812621)


## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Flavor