Orange Lamps (#1456)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Adds a new street lamp variant that is based on the green one, but with
the orange light of the regular lamptern one. It also adds a wall
version of the new orange lamp and changes the lamp posts within the
keep to the new orange ones.

## Why It's Good For The Game

The green street lamp is the better of the two designs, and better fits
the new renaissance theme of the server, but the green light is freaky,
way too alien and an eyesore to look at. A more normal, orange light
version is the best of both worlds between keeping that design and
having a light that is not painful to look at.

New orange lamp:

![lamp1](https://github.com/user-attachments/assets/9de66308-0a21-4cce-8622-cb1784522289)

Lamp when it's off:

![lampoff](https://github.com/user-attachments/assets/fb6daf18-237e-40bf-80ba-24a02466caf1)

New orange midlamp:

![lampmid](https://github.com/user-attachments/assets/1a07d68e-463f-4573-8e79-f2ef1b70d905)

And the brand new wall lamp:

![wallamp](https://github.com/user-attachments/assets/f4eda8bd-a654-4463-890d-c80b174f406b)