Now old Court magicans can scry (#1458)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Fixes the if switch using an "or" instead of a "to" to include level 5
and 6 of arcana.


![magican](https://github.com/user-attachments/assets/2e8e26ba-7204-4b58-9afa-6acf52c70e18)

![Scrying](https://github.com/user-attachments/assets/08f5d290-17fb-494a-b76a-f953d309ecfd)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
fixes #1402
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->