Fixes paralyzed mobs from being able to resist (#1460)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds a check to see if the mob is paralized before they can attempt to
resist.

![image](https://github.com/user-attachments/assets/7278fc78-2f23-42e0-83dc-aa7e2a6a9373)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
fixes #954
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->