Lamp Post Layer Fix (#1462)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Fixes the layering of the lamp posts so people aren't standing on top of
it, and removes some of the lamp posts around the city to make some
parts darker.

## Why It's Good For The Game

People are, tragically, no longer standing atop of the lamp posts. This
looks better. Alleyways will now also be darker, allowing for more
Zizoite, Matthiosian (and Baothan) heresy to occur away from prying
eyes...


![Untitled](https://github.com/user-attachments/assets/ea7a282d-525d-40ac-9322-0ca2fddf524a)
_Where'd the duke go?!_