Lightens dark gloves (#1464)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Makes black gloves not pitch black
old

![image](https://github.com/user-attachments/assets/2fa1a678-c3bd-48e2-9e87-350bf69af3ce)
new

![image](https://github.com/user-attachments/assets/41d46f24-83ff-4f46-a5ee-cf9f0a2b7190)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Better sprite is good for the game
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->