Make both soilson and soilbride get keys (#1475)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Moves the repeated things be before the if, including the key which the
soilbride lacked.


![image](https://github.com/user-attachments/assets/8c2a3998-7738-4d2c-ac70-62983894190e)

![works](https://github.com/user-attachments/assets/a9a92930-10ae-4b40-879f-af65094bcc37)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
I see no reason for a Soilbride to not have keys.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->