Mercenary Adjustments 2.whatever (#1469)
## About The Pull Request
Adjusts the mercenary guild's rooftop overhang, so it can't be climbed
onto as easily, if at all, from prior locations. Without building
upwards, anyhow.

Adjusts the Kommandant's PQ requirement to be 6 and gives him some
better equipment to differentiate him easier. Blacksteel chausses and
blacksteel boots. Also gives him a chainmail shirt instead of the poofy
one, and a chainmail coif.

Removes one slot from Grenzel, just to account for the Kommandant, even
though we'd never have more than ten slots in total. Just because it was
bugging me.

## Why It's Good For The Game
Kommandant is intended to be a better equipped role, with a higher
standard of play requirement. This reflects that, by upping the PQ
requirement and significantly adjusting their equipment.

As for the rooftop, it's just because I like to climb on it, smash
windows and flee. The players didn't like this. :(