Court Physician's Garden Overhang (#1480)
## About The Pull Request
Adds an overhang above the Court Physician's garden. It can still get
sunlight, but it can't be dropped onto. Should prevent someone from
looting his facilities the easy way, round start, and force some new
creativity when getting into the keep.

## Why It's Good For The Game
For the same reason we can't have nice things elsewhere. Because of me.
More importantly, it's because folks would get up on the keep's walls
via trees, the gate, whatever, and then drop into the Court Physician's
chambers. They'd steal his red, the organs, spare bottles, etc. This is
obviously not kosher, because it can happen within a few minutes of the
round starting and prevent literally any play on his part.
I abhor hugbox'd things, but this might be one of the few things I agree
to do this for.

Is this objective? No. Absolutely not. Some people like map enforced
trauma. While I'd typically agree, I think it's overplayed in this case.