TEMPORARY wooden fork icon placeholder (#1490)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Makes the wooden forks use the iron fork sprite **TEMPORARILY**, i've
tried searching the other repositories and they are also lacking the
same icon, this should make the forks have icons **TEMPORARILY** until
someone makes a proper icon or finds the one that was deleted.
Before:
![Captura de pantalla 2024-11-07
204410](https://github.com/user-attachments/assets/29e56c2c-0f39-4c3d-8858-edcae82de3f0)
After:
![Captura de pantalla 2024-11-07
204325](https://github.com/user-attachments/assets/c94e9b78-198b-4e46-9be9-202a33a91808)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

Probably isn't going to change anything, but i like my things with
icons, even if it's a placeholder for now
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->