Logs Seelie wing ripping. (#1497)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Now if a Seelie get's their wings ripped it get's logged in both the
Ripper and the Seelie.

![image](https://github.com/user-attachments/assets/19726731-1782-45e3-839a-00254836994d)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Helps the Admins see events better.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->