Tiny admin verb nitpick - "Aghost (Toggle Invisibility)" (#1501)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Currently this is the verb with the longest name in the GameMaster tab.
I would rather it not be so.

Also, it's a little bit weird that this isn't a preference, but actually
a `client` var that doesn't survive reconnects.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Makes it not force you to adjust the chat screen so far to the left if
you want 3 verbs per row
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->