Gives a price to aged cheese (#1491)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
 
![Captura de pantalla 2024-11-07
204824](https://github.com/user-attachments/assets/9ce6a48d-8525-4159-979b-9a36f9759bf0)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

More ways to make money! Now cheese can be let to age and exported on
the merchant's balloon for a decent amount of mammon, taking into
consideration the amount of money cheese costs to buy and the amount of
time it takes to age.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->