Runtime fix for fog walls + show fuzzy time remaining to bandits (#1499)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
*With apologies to Carl.*

- Fixes the insane amounts of runtimes caused by fogwalls trying to
access the `mind`s of `mind`-less mobs *everywhere*

And, because I always get carried away:
- Prevents the fog wall from spamming your chat every 0.5s you walk into
it (now: 40-second cooldown. Uses a timer, a global list, and a proc to
remove the mob from that list.)
- Makes the warning message less unnecessarily direct and IC-ly weird.
And more informative (vaguely) as to how long before the barriers go
down!
- Examine will also give you a vague description of remaining time.

Future suggestions: perhaps make a single global timer that gets added
on round start, and asynchronously delete all the walls when it
triggers, instead of having 346 different 30-minute timers. Likely not
going to be a huge performance diff, but it would be cleaner.
Having the timers start on round start, rather than when initialized,
would also make the walls *actually* last 30 minutes, instead of 30m
minus however long round setup took. CBA to figure that out now because
I've spent enough time on this sideproject.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
![Screenshot from 2024-11-07
16-06-38](https://github.com/user-attachments/assets/30098fc3-0527-4d07-9e2c-12cc20c92619)
![Screenshot from 2024-11-07
17-15-47](https://github.com/user-attachments/assets/bd787f36-832b-4ed8-a71a-e9784a887088)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->