[TM first] Light Step Changes & New SPECIAL (#1467)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
This was literally just /tg/ light step. It hadn't even been recoded to
interact with any of our new systems, including the FOV cone.

As it stands; this now also adds a new SPECIAL - 'Underworld Assassin'
-Takes -1 STR
-Gives 2 END
-Gives 2 SPD
-Gives 2 SNEAKING
-Gives 1 CLIMBING & ATHLETICS
-Light step

This now goes ahead and makes Light Step into Lightfooted, a rarer
version (with the original removed) that does three primary things;

-Entirely Nullifies being able to see people moving behind you when they
have this trait
-Removes the sound of their footsteps
-Removes the sound of their armor clanking (This is likely going to be
nerfed if its _too good_.
+1 SPEED and ATHLETICS
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This was one of the most dogshit additions to the SPECIAL PR. Ever.
(involving the original addition of the light step trait). This
hopefully rounds it out, and makes it better.

It was a trait that straight up just used /tg/ content with 0 changes, 0
additions, 0 care.

This makes it actually do something.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->