Lightning bolt nerf/adjustment (#1516)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Lightning bolt no longer hard stuns. Instead, it causes a second long
paralysis, similar to repulse.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This should be much less terrible feeling for people on the recieving
end, while still being rather useful. A ranged knockdown that allows a
mage a bit of breathing room to cast another spell, that isn't a hitscan
fight ender. It should also be much harder to chain murder people with
it and fireball now.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->