Fixes weird grass tiling. (#1513)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Before:
![Screenshot 2024-11-08
191617](https://github.com/user-attachments/assets/c89e978c-e556-4f5b-8988-95de159a122a)
After:

![image](https://github.com/user-attachments/assets/94ebfa96-d632-4967-acb2-86a1797fb36a)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This one bugged me for a long time. 
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>