Patches some 64x64 weapon sprites. (#1511)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Couple of them needed some touch-up, so that is what I have done!

![image](https://github.com/user-attachments/assets/59f88a8c-554a-4526-a2c3-0c1dd178ac09)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Brings them to our metal palettes + better shading. 
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>