Artificer Expansion Pack: Contraptions (#1479)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This Expansion Pack (29.99$) includes two new toys!
CONPTRAPTIONS!

Now Artificers get to make a machine and experiment with it!

These contraptions require something to power them, which wastes away
after X amount of uses (defined by contraption).
In what you'll see in these videos it them taking cogs to function. This
will probably change in the future at least for the smelter (it will
accept coal instead).

These machines have a chance to misfire, as you will see in a video down
there, this chance is reduced with engineering skill and good luck!

The idea here is to make machines which have creative uses, that are
slightly comical, fun to play around with and are unpredictable.
They hopefully reward experimentation having very different outcomes and
things that may happen to them!
A very important part of this is the feel and sound of them. Which I
took some care with!

They were made with future machines in mind, making it easy to code
more!

Presenting the contraptions:

***At present I haven't made sprites for them yet!***

### Portable Smelter

This handy device will turn things into other things.
Yeah, it will smelt them, go figure.

***This contraption has 6 uses and a misfire chance of 10.***

https://www.youtube.com/watch?v=YJiwSwH09yg&feature=youtu.be

I've gone the extra mile and added smelt results to some other unusual
items, where it would make sense.

### Wood Metalizer

This one is funky.
Its a machine that turns wood into metal. What the fuuuuuuuuuck?!

***This contraption has 5 uses (in the video you see 3) and a misfire
chance of 10.***

https://www.youtube.com/watch?v=0X1RaJEx6sc

#### Now, its important to say these machines have a chance to fail in a
comical way.

Here's me fucking around and finding out:

https://www.youtube.com/watch?v=BcHQlUjmJSs

## Why It's Good For The Game

I was asked like 3 times what the Artificer was about after making the
last PR, and honestly, I couldn't give a straight answer.
My vision is to make the artificers have wacky insane machines like
these that reward experimentation.

The theme of this pack is metal because that's the other half of
Artificing that I didn't really touch and is a good way to ensure
artificers and smiths love each other in a weird way.

## Changelog

add: Added two contraptions, the portable smelter and the wood metalizer
tweak: Some items were previously unable to be clicked or attacked. Now
they can. (Think of sticks)
code: Moved the smeltresult var from obj/item to obj