Hotfix for Burial Changes - the AFTER LIFE button. (#1521)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
AFTER LIFE completely bypasses the changes to burial. Not intentional. I
cleaned up the button's Click() code, removing useless commented-out
stuff and currently-unused features. I also moved the logic of the
"Journey to the Underworld?" proc to another client proc, both for
consistency and to avoid copypasting code.
This also fixes a runtime related to your verbs not being removed -
after making the underworld spirit, the AFTER LIFE button tries to
remove the ghost verbs... from the ghost's client, which is null,
because the client already got transferred to the spirit.

Additionally, if you're funeralized: you will be warned about having
been buried, and it's hinted that you'll be sent straight to the lobby.
(You'll also get the message afterwards explicitly telling you as much.)
The prompt echoes that of the underworld carriage.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
Oversight is fixed. Runtime is fixed. People get clued in on whether
they respawn or not.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

---------

Co-authored-by: zoktiik <zoktiik@users.noreply.github.com>