Blacklists Hardened Leather armor from Masterwork Recipe (#1524)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Hardened leather was able to be used in crafting the masterwork armor.

Masterwork was intended to be an upgrade to hunter leather armor, not
the fucking be all end all of the universe.

This solves that by blacklisting the one armor Seamstresses can make
from the masterwork recipe.

## Why It's Good For The Game

Balance, fuck, and fix..