Artificer Expansion: Lock Imprinter (#1519)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

This adds yet another Contraption!

Presenting...

### The Lock Imprinter!


![image](https://github.com/user-attachments/assets/364f6e3e-b5e0-4ba3-89e9-cbaee6caa0cd)

The lock imprinter is a hard to craft artificer contraption that focuses
around locks!


![image](https://github.com/user-attachments/assets/400d47d0-f953-4eab-88a3-a6c5391c88bf)

It has three modes, Examine, Imprint and Unlock.

It will Examine lock IDs.
Imprint a lock ID into something that has a lock.
Lock and unlock something.

https://www.youtube.com/watch?v=EygnZTvDB0w

Now, I know this might be a taaaaaad scary.

But I did my best to ensure this was balanced because:

1 - Low charge count (2) and peculiar charge item (lock)
2 - Engineering skill 5 required (it cannot be made within the first 40
minutes of a round at minimum)
3 - It has a chance to explode in the hands of those without an
engineering skill (this chance is hidden except from a legendary
artificer)
4 - Usage is very loud, animated and displays a warning on chat to all
surrounding players (which can't be seen in the video since I hadn't
added that part yet)
5 - It requires either an extreme bout of luck, time wasted leveling up
lumberjack skill or a wood cutter to be in the round and offer the
necessary Essence of Lumber.

The idea here is that it could be used for good and bad!

You help people lock their their belongings away, fix locks and reset
locks in things like doors at request!
Or you can be a gremlin and either give it to some beggar and watch him
explode himself or become a little thiefing rat.

### Misc changes!

Here are some misc changes!

#### Wooden Organs!

Just kidding, all you get is a fucking wooden eyeball.
I guess it may have some use?!

#### Smelter and Wood Metalizer now have special misfires

I wont tell everything because it spoils the fun, but now misfiring
should be a little more unpredictable and "wtf" but also may be useful
under extremely specific circunstances.

#### Moved Crowsbow and Ammo to artificer

I believe it's good to ease the burden from smithing a bit by giving all
ranged options to the artificer instead.
Specially because they are made of wood and thats their thang.

#### Expanded Metalizer results a bit

Gave Metalizer results to items that didn't have them.
Changed some metalizer results from items the Artificer can craft to not
only make more sense but also be more compatible with recipes and
contraptions

#### Technically made the Runelock and Runelock Bullets obtainable

Now, I say technically because its not THAT direct.
I'm not saying it here (I don't want to spoil the fun) but you can
figure it out in the code!

#### Slightly Improved contraption code and artificer recipe code

Mainly the recipes now allow for multiple results.

## Why It's Good For The Game

I had a really good time playing Artificer with another playing, going
on a journey of figuring out what interacted with what and what
contraptions actually did.
Hopefully this content further strengthens the Artificer role and gives
it a very special niche as the champion for SCIENCE, EXPERIMENTATION and
INNOVATION!

Noc and Malum be praised!

## PS

I had to fucking change a lot of "var/lockid" vars into "lockid".
I just moved var/lockid to fucking object instead, did the same with
"var/locked" and "var/hashlock".
These are very important for the Contraptions to work and to keep them
modular and in no way matter or change anything anywhere else.

I also had to standardize the way lockids work, there were like 3 types
of lock id.
2 I in vars were lockid and keycontrol.
I renamed keycontrol to lockid, everything works beautifully as can be
seen in the video!