Gives the butcher a key ring. (#1528)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Now the butcher starts with both the butcher and farm keys.

![image](https://github.com/user-attachments/assets/935eb5d7-ac65-4726-a1c7-c3a8fa496453)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game
This allows the butcher to enter the farm area.
<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->