Remove all but one Purity machines, located in a crackden. Dramatically increases drug prices. (#1505)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
Title.


![image](https://github.com/user-attachments/assets/fb4bf0e8-10b1-463c-925f-d04e4b74fb86)


## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->
Cheap, safe and readily available drugs are ruining the economy.
Drug machines in the keep, the town and the bog have turned Rockhill in
the most progressive city in the continent.
NOURISH especially has rendered most foodmaking useless.