Removes artificer keys from woodcutters (#1527)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

Tittle

## Why It's Good For The Game

The number of people who have keys for the place is too high, the place
has a grand furnace, and artificers are supposed to sleep inside and
keep their things, but you can't be safe with so many people who can
just walk in. Woodcutters feel entitled to getting wood planks; instead
of just dumping the logs in the stockpile they put the planks, I've
noticed that a lot of times there are no logs in the stockpile while
there are dozens of planks inside.