Citywatch Armour Pass (#1531)
## About The Pull Request
Gives the citywatch helmets a higher integrity by 25, smash protection
from crits and ups their armour class accordingly. I figured it was
supposed to be more 'riot' based. Their armour got similar treatment,
receiving smash protection from crits. Additionally, it now covers the
neck, as it visibly has a neck guard. Also assumedly meant to be more
for 'riot control', and as such, I thought it fit.

Gives a description to both items.

Introduces crafting recipes for both items. 2 steel / 1 coal for the
helm, 3 steel / 1 coal for the armour.
## Why It's Good For The Game
Citywatch used to spawn with a coif. This replaces that without making
the armour look worse by an addition to their neck slot. Also makes
sense logically. Additionally, it made sense to have it be resistant to
more blunt damage, like smashing, since they fight townsfolk and the
like.

Upping the durability on the helmet is more a personal choice at the
request of a few folks, because apparently it's super easy to break as
is.

Crafting recipes for the sake of them not being able to get it back
after it's broken.