Artificer Small QOL Changes (#1530)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request

- Fixed Wood Metalizer interaction with Crossbow (for real this time)
Proof I didnt fuck crosbows up again:


![image](https://github.com/user-attachments/assets/ddee08bd-f2aa-48f8-9ef0-4793e7ab1faf)

- Metal cogs can now also be made from Copper and Tin
Cogs are a basic necessity for artificer. No sense in making them so
limited to an ALLOY.

- Deducted 1 plank cost form certain basic recipes
These were eating away at wood and time for no real reason and mostly
happened as an oversight on my part.

- Imprinter no longer requires level 5 engineering to build.
I was probably overzealous when balancing the imprinter as I thought it
would be extremely strong.
Not only is it slept on but there's already a soft time gate in the form
of the Essence of Lumber required to create it.

- Contraptions no longer misfire in the hands of Legendary Artificers
How I didn't think of this before is astonishing.
The misfires are fun if they're directly connected to some ineptitude.
It always being up to chance wasn't conveying the feeling I wanted it
to.

- Added Wood Metalizer result for wooden windows
Now you get to reinforce windows the same way you doors... and smelt
them all the same...

- Added smelt result to bars
It was very annoying when after exploding for the firth time as
artificer you'd inevitably destroy a table. But the bars would remain.
Essentially disallowing you to do anything with that tile. Bars begone.

## Why It's Good For The Game

These are some QOL and balance changes for Artificers.
They will benefit how the role is played!