Adds bullet pouches and refactored Quiver code. (#1537)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->

## About The Pull Request
Adds bullet pouches and refactored the quiver code to accept future ammo
storage options, so now the Quiver is a subtype of `ammo_storage/`.

**Do note that they are using the pouch sprites and are not craftable.**

![pouch_e](https://github.com/user-attachments/assets/96a9a5f5-0778-4714-abe2-50b821201fb2)



https://github.com/user-attachments/assets/267fd60d-685a-4db4-8e97-4b3dca6349ad


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

## Why It's Good For The Game

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

Makes carrying ammo for these guns less painful.