Avatar of Dendor
Redoes the volf pit. Lowers it a Z-Level and replaces the volves with a singular Verewolf that's been cracked out. A lesson from Kingsfield.
To get out of the pit now, you need the gate open to climb out, alongside needing a high enough climbing skill. Someone could probably throw down a chair, but I 'unno if that'd be enough for someone without climbing skill in general.

I found this funny. That's all. More interesting that three volves.