Watchmen Armoury Access (#1544)
## About The Pull Request
Removes a copy of the Sheriff's keychain for some reason and gives
armoury access to the Watchmen.

## Why It's Good For The Game
They bash down the door anyways, and there's rarely a Sheriff. While the
former isn't reason to give them keys, I don't believe this would be too
big a change. It just gives them access to an area they otherwise
couldn't make use of in a round normally, alongside letting them make
use of the skills they have.