Prisoner Hermes (#1546)
## About The Pull Request
Adds Hermes to prisoner and hostage cells within Rockhill's town.
Doesn't do the same for the Gate's(Bog) prisoners, as they're assumed to
be more frequently interacted with.

Also corrects the Mercenary Guild's hermes tag, so they're not just
marked as 'inside'.

## Why It's Good For The Game
This was at request. Has people more involved in the round actively if
they're ignored, able to mail folks.