Ports Ratwood's feint buffs (INT scaling, minor SPD scaling, feinted debuff duration increase)
Feint success probability now scales with INT (+-10% for each point of difference) properly instead of being flat +15% for having more INT than your opponent, be it 1 point difference, or 10. 

Minor scaling with SPD (+-5% for each point of difference). Caps feint probability at 90%, standard for all combat procs. 

Duration of applied debuffs is increased only if you are using feint mode. It is still possible to use feints in any mode, but with much of the effect lost.